% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny_coda.R
\name{array_munge}
\alias{array_munge}
\alias{array_pre}
\alias{array_post}
\title{Functions for setting up transforms and handing weird arrays}
\usage{
array_pre(X, b)

array_post(X, b, s)
}
\arguments{
\item{X}{array}

\item{b}{part/coord index in array}
}
\value{
matrix (if array_pre) array (if array_post)

s dimensions of array desired (for array_post)
}
\description{
Functions for setting up transforms and handing weird arrays
}
