% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{base_lr_transforms}
\alias{base_lr_transforms}
\alias{alrContrast}
\alias{clrContrast}
\alias{ilrContrast}
\alias{glr}
\alias{glrInv}
\alias{alr}
\alias{alrInv}
\alias{clr}
\alias{clrInv}
\alias{ilr}
\alias{ilrInv}
\title{Log-Ratio Transformation}
\usage{
alrContrast(D, d, inv)

clrContrast(D, inv)

ilrContrast(D)

glr(X, V)

glrInv(X, V)

alr(X, d)

alrInv(X, d)

clr(X)

clrInv(X)

ilr(X, V = NULL)

ilrInv(X, V = NULL)
}
\arguments{
\item{D}{the number of parts (e.g., number of columns in untransformed data)}

\item{d}{for ALR, which component (integer position) to take as reference
(default is nrow(x)) for alrInv corresponds to column position in untransformed
matrix.}

\item{inv}{for ALR and CLR, transformation matrix is different forward and inverse}

\item{X}{vector or matrix (rows are parts/coordinates, columns are samples) of data}

\item{V}{transformation matrix (defines transform; P x D where D is number of parts)
if NULL then uses ilr default basis (see details)}
}
\value{
matrix (converts vectors to column matricies)
}
\description{
\code{glr} is generic log-ratio transform, code used by other
transforms, can be called directly. *Contrast functions produce contrast 
matricies (e.g., "V") that define the various transforms.
}
\details{
The default ILR base formed by Gram-Schmidt orthogonalization of an ALR_D basis.
}
\examples{
#ALR Transform
x <- matrix(runif(30), 10, 3)
x <- clo(x)
x.alr <- alr(x, 2)
x <- alrInv(x.alr, 2)

# ILR
x.ilr <- ilr(x)
x <- ilrInv(x.ilr)

# CLR
x.clr <- clr(x)
x <- clrInv(x.clr)

# CUSTOM - Be careful if your custom matrix is not
# orthogonal the inverse transform may not be given by just the transpose!
# For example, this is the case for the ALR
V <- matrix(c(1, 1, -1), 1, 3)
x.custom <- glr(x, V)
}
