% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{center}
\alias{center}
\title{Center - subtract each column by column mean}
\usage{
center(X)
}
\arguments{
\item{X}{matrix (D x N)}
}
\value{
matrix of same dimension as X
}
\description{
Center - subtract each column by column mean
}
