% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clo}
\alias{clo}
\title{Closure divide each column by column sum}
\usage{
clo(X)
}
\arguments{
\item{X}{matrix (D x N)}
}
\value{
matrix of same dimension as X
}
\description{
Closure divide each column by column sum
}
