% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_coda.R
\name{clrvar2phi}
\alias{clrvar2phi}
\title{Calculate Phi statistics (proportionality) from CLR Covariances}
\usage{
clrvar2phi(Sigma)
}
\arguments{
\item{Sigma}{Covariance matrix Px(PN) where N is number of 
covariance matricies in CLR space}
}
\value{
Array (same dimension as Sigma) but elements represent phi statistics
}
\description{
Assumes parts are the first two dimensions of Sigma
}
\references{
Lovell, David, Vera Pawlowsky-Glahn, Juan Jose Egozcue, 
  Samuel Marguerat, and Jurg Bahler. 2015. Proportionality: A Valid Alternative 
  to Correlation for Relative Data. PLoS Computational Biology 11 (3). 
  Public Library of Science: e1004075.
}
