% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/transfer_coda.R
\name{iiContrast}
\alias{iiContrast}
\alias{icContrast}
\alias{ciContrast}
\alias{iaContrast}
\alias{aiContrast}
\alias{caContrast}
\alias{acContrast}
\alias{aaContrast}
\alias{convert_coda}
\alias{ilr2ilr}
\alias{ilr2clr}
\alias{clr2ilr}
\alias{alr2clr}
\alias{clr2alr}
\alias{alr2alr}
\alias{ilr2alr}
\alias{alr2ilr}
\alias{ilrvar2ilrvar}
\alias{ilrvar2clrvar}
\alias{clrvar2ilrvar}
\alias{alrvar2clrvar}
\alias{clrvar2alrvar}
\alias{alrvar2alrvar}
\alias{ilrvar2alrvar}
\alias{alrvar2ilrvar}
\title{Transfer data and covariance matricies from one coordinate system to another}
\usage{
iiContrast(V1, V2)

icContrast(V1)

ciContrast(V2)

iaContrast(V1, d2, D)

aiContrast(d2, V2, D)

caContrast(d2, D)

acContrast(d1, D)

aaContrast(d1, d2, D)

ilr2ilr(X, V1, V2, b = 1)

ilr2clr(X, V1, b = 1)

clr2ilr(X, V2, b = 1)

alr2clr(X, d1, b = 1)

clr2alr(X, d2, b = 1)

alr2alr(X, d1, d2, b = 1)

ilr2alr(X, V1, d2, b = 1)

alr2ilr(X, d1, V2, b = 1)

ilrvar2ilrvar(Sigma, V1, V2)

ilrvar2clrvar(Sigma, V1)

clrvar2ilrvar(Sigma, V2)

alrvar2clrvar(Sigma, d1)

clrvar2alrvar(Sigma, d2)

alrvar2alrvar(Sigma, d1, d2)

ilrvar2alrvar(Sigma, V1, d2)

alrvar2ilrvar(Sigma, d1, V2)
}
\arguments{
\item{V1}{ILR contrast matrix of basis Sigma is already in}

\item{V2}{ILR contrast matrix of basis Sigma is desired in}

\item{d2}{alr reference element Sigma is to be expressed with respect to}

\item{d1}{alr reference element Sigma is already expressed with respec to}

\item{X}{data to transform between representations}

\item{b}{index of dimension to operate on (e.g., index of dimension of parts or coords in X;
default is 1 meaning that compositions/log-ratios are rows) Note, for covariance matricies, 
b is meaningless, first two dimensions must be log-ratio coordinates.}

\item{Sigma}{covariance matrix in specified transformed space}

\item{V}{ILR contrast matrix (i.e., transformation matrix of ILR)}
}
\value{
matrix
}
\description{
Naming Conventions:
xxContrast produces a contrast matrix for transforming between representations where 
i=ILR, c=CLR, a=ALR. x2x represent transformations of data between coordinate systems 
xvar2xvar represent transformations of covariance matricies between coordinate systems. 
By convention all functions take arguments as follows: all functions have the form function(a,b,c)
where a is the data/covariance, b is the argument for specifying the current coordinate system, and c 
represents the argument for specifying the desired coordinate system.
}
\examples{
# For Data
X <- matrix(abs(rnorm(10)), 5, 2)
X <- clo(X)
X.clr <- clr(X)
X.alr <- clr2alr(X, 3)

# For a covariance matrix starting in the alr
Sigma.alr <- X.alr\%*\%t(X.alr) # Just a random covariance matrix 
Sigma.ilr <- alrvar2ilrvar(Sigma.alr, 3, ilrContrast(5)) # Covert to default ILR

# Another way of doing this more "manually" - use contrast matricies
V <- caContrast(3, nrow(X)) # create contrast from clr to alr
X.alr <- V \%*\% X.clr
}
