% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/transfer_coda.R
\name{iqlr}
\alias{iqlr}
\alias{var2iqlrContrast}
\alias{clrvar2iqlrvar}
\title{Functions for implementing / calculating IQLR}
\usage{
var2iqlrContrast(S, qLow = 0.25, qHigh = 0.75)

clrvar2iqlrvar(Sigma, qLow = 0.25, qHigh = 0.75)
}
\arguments{
\item{S}{P-vector of "variances" for IQLR (e.g., diagonal of CLR covariance matrix)}

\item{qLow}{lower quantile (between 0 and 1)}

\item{qHigh}{upper quantile (between 0 and 1; must be greater than qLow)}

\item{Sigma}{(PN)xP covariance matrix in CLR}
}
\value{
PxP contrast matrix (var2iqlrContrast) or 
   (PN)xP covariance matricies (clrvar2iqlrvar)
}
\description{
var2iqlrContrast calculates contrast matrix in iqlr based on part-wise variance 
and a specification for the lower and upper quantiles to use.
}
\details{
clrvar2iqlrvar transforms covariance matrices into the iqlr. Note a different 
iqlr is calculated for each covariance matrix

These are advanced features and should not be used without prior knowledge of the
IQLR transform.
}
\references{
Jia R Wu, Jean M Macklaim, Briana L Genge, Gregory B Gloor. 
  Finding the centre: corrections for asymmetry in high-throughput sequencing datasets. 
  (2017) arXiv preprint arXiv:1704.01841
}
