% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codamath.R
\name{linForm}
\alias{linForm}
\title{Linear form Y=VX}
\usage{
linForm(X, V1, V2 = NULL, b = 1)
}
\arguments{
\item{X}{vector, matrix, or array of data (b is the dimension on which the linear form acts 
e.g., the dimension on which V1 and V2 are multiplied).}

\item{V1}{(P1 X D1, e.g., transfer Contrast matrix)}

\item{V2}{(P2 x D2, e.g., transfer Contrast matrix)}

\item{b}{(index of dimension to operate on; default: 1)}
}
\value{
Y
}
\description{
Linear form Y=VX
}
\details{
Calculates

  Y = bdiag[V1, V2]  X
where bdiag a block diagonal matrix with blocks V1 and V2.
If V2 is not given then assumes that second block is identity.
In that case determines the size of second block based on size of 1.
}
