% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codamath.R
\name{quadForm}
\alias{quadForm}
\title{Quadratic Form Y = V X V^T}
\usage{
quadForm(X, V1, V2 = NULL)
}
\arguments{
\item{X}{matrix or array (If array, assumes first two dimensions 
are the ones to operate on, e.g., X_i = X[,,i] if 3D array)}

\item{V1}{(e.g., transfer Contrast matrix)}

\item{V2}{(e.g., transfer Contrast matrix)}
}
\value{
Y
}
\description{
Quadratic Form Y = V X V^T
}
\details{
Calculates:

Y_i = bdiag[V1, V2]  X_i  bdiag[V1, V2]^T

where Y_i and X_i refer to the i-th set of D-columns of Y and X respectively
and where bdiag a block diagonal matrix with blocks V1 and V2.
If V2 is not given then assumes that second block is identity.
In that case determines the size of second block based on size of V1.
}
