% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCoDA.R
\docType{package}
\name{RcppCoDA}
\alias{RcppCoDA}
\alias{RcppCoDA-package}
\title{RcppCoDA}
\description{
Really Fast implementation of Compositional Data Analysis  (CoDA) transforms and 
functions using C++ and accelerated by the Eigen Library. Major functions
are implemented as header only library for import into other packages.
}
\author{
Justin D Silverman
}
