% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny_coda.R
\name{vec_to_array}
\alias{vec_to_array}
\title{If X is a vector - convert it into a matrix (with 1 column)
If X is an array it stays an array}
\usage{
vec_to_array(x)
}
\arguments{
\item{x}{thing to test/transform}
}
\value{
matrix
}
\description{
If X is a vector - convert it into a matrix (with 1 column)
If X is an array it stays an array
}
\examples{
vec_to_array(c(1,2,3))
vec_to_array(rbind(c(1,2,3), c(1,2,3)))
}
