% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{center}
\alias{center}
\title{Center operation - subtract from each element the mean of elements in b dimension}
\usage{
center(X)
}
\arguments{
\item{X}{vector, matrix, or array of data
(b is the dimension which is relevant to be centered)}

\item{b}{index of dimension to operate on
(e.g., index of dimension of parts or coords in X;
default is 1 meaning that compositions/log-ratios are rows)}
}
\description{
Center operation - subtract from each element the mean of elements in b dimension
}
