% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clo}
\alias{clo}
\title{Closure operation - divide elements by sum of elements in b dimension}
\usage{
clo(X)
}
\arguments{
\item{X}{vector, matrix, or array of data
(b is the dimension which is relevant to be closed)}

\item{b}{index of dimension to operate on
(e.g., index of dimension of parts or coords in X;
default is 1 meaning that compositions/log-ratios are rows)}
}
\description{
Closure operation - divide elements by sum of elements in b dimension
}
